requirements.txt
pytest==8.2.0
assertpy==1.1
docker==7.1.0
allure-pytest==2.13.5